Update add product fragment business logic and refactor code

+++ app/src/main/java/com/app/stockmanagement/data/local/entity/ProductEntity.kt
@@ -12,7 +12,7 @@ data class ProductEntity(
     val price: Double,
     val category: String,
     val barcode: String,
-    val supplierId: Int,
+    val supplierId: Long,
     val currentStockLevel: Int,
     val minimumStockLevel: Int
 )
