+++ app/src/main/java/com/app/stockmanagement/domain/model/Product.kt
@@ -6,7 +6,7 @@ data class Product(
     val price: Double,
     val category: String,
     val barcode: String,
-    val supplierId: Int,
+    val supplierId: Long,
     val currentStockLevel: Int,
     val minimumStockLevel: Int
 )
